# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BuildRMC Fortune Dashboard API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [billing]
      summary: List all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [billing]
      summary: Get an invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    put:
      operationId: updateInvoice
      tags: [billing]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [billing]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /employees:
    get:
      operationId: getEmployees
      tags: [hrm]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [hrm]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [hrm]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [hrm]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [hrm]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /attendance:
    get:
      operationId: getAttendanceRecords
      tags: [hrm]
      summary: List attendance records
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendanceRecord
      tags: [hrm]
      summary: Create an attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRecordBody"
      responses:
        "201":
          description: Attendance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /payroll:
    get:
      operationId: getPayrollRecords
      tags: [hrm]
      summary: List payroll records
      responses:
        "200":
          description: List of payroll records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRecord"
    post:
      operationId: createPayrollRecord
      tags: [hrm]
      summary: Create a payroll record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayrollRecordBody"
      responses:
        "201":
          description: Payroll record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRecord"
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - fullName
        - email
        - role
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        contact:
          type: string
        email:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        creditTerms:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - address
        - contact
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        contact:
          type: string
        email:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        creditTerms:
          type: ["string", "null"]
      required:
        - name
        - address
        - contact
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        plant:
          type: ["string", "null"]
        loadedPlant:
          type: ["string", "null"]
        kmReading:
          type: ["number", "null"]
        block:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        invoiceTime:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        vehicleId:
          type: ["integer", "null"]
        vehicleNo:
          type: ["string", "null"]
        pumpType:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        loadedGrade:
          type: ["string", "null"]
        loadedQuantity:
          type: ["number", "null"]
        quantity:
          type: ["number", "null"]
        netAmount:
          type: ["number", "null"]
        netPrice:
          type: ["number", "null"]
        pumpCharge:
          type: ["number", "null"]
        transportCharge:
          type: ["number", "null"]
        cgstRate:
          type: ["number", "null"]
        sgstRate:
          type: ["number", "null"]
        igstRate:
          type: ["number", "null"]
        remark:
          type: ["string", "null"]
        isBillReceived:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - invoiceDate
        - customerId
        - customerName
        - totalAmount
        - status
        - isBillReceived
        - createdAt
    CreateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        customerId:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        plant:
          type: ["string", "null"]
        loadedPlant:
          type: ["string", "null"]
        kmReading:
          type: ["number", "null"]
        block:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        invoiceTime:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        vehicleId:
          type: ["integer", "null"]
        vehicleNo:
          type: ["string", "null"]
        pumpType:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        loadedGrade:
          type: ["string", "null"]
        loadedQuantity:
          type: ["number", "null"]
        quantity:
          type: ["number", "null"]
        netAmount:
          type: ["number", "null"]
        netPrice:
          type: ["number", "null"]
        pumpCharge:
          type: ["number", "null"]
        transportCharge:
          type: ["number", "null"]
        cgstRate:
          type: ["number", "null"]
        sgstRate:
          type: ["number", "null"]
        igstRate:
          type: ["number", "null"]
        remark:
          type: ["string", "null"]
        isBillReceived:
          type: boolean
      required:
        - invoiceNumber
        - invoiceDate
        - customerId
        - totalAmount
        - status
    UpdateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        customerId:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        plant:
          type: ["string", "null"]
        loadedPlant:
          type: ["string", "null"]
        kmReading:
          type: ["number", "null"]
        block:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        invoiceTime:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        vehicleId:
          type: ["integer", "null"]
        vehicleNo:
          type: ["string", "null"]
        pumpType:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        loadedGrade:
          type: ["string", "null"]
        loadedQuantity:
          type: ["number", "null"]
        quantity:
          type: ["number", "null"]
        netAmount:
          type: ["number", "null"]
        netPrice:
          type: ["number", "null"]
        pumpCharge:
          type: ["number", "null"]
        transportCharge:
          type: ["number", "null"]
        cgstRate:
          type: ["number", "null"]
        sgstRate:
          type: ["number", "null"]
        igstRate:
          type: ["number", "null"]
        remark:
          type: ["string", "null"]
        isBillReceived:
          type: boolean
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        department:
          type: string
        designation:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        salary:
          type: number
        joinDate:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - department
        - designation
        - salary
        - joinDate
        - status
        - createdAt
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        designation:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        salary:
          type: number
        joinDate:
          type: string
        status:
          type: string
      required:
        - name
        - department
        - designation
        - salary
        - joinDate
        - status
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        date:
          type: string
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - employeeName
        - date
        - status
        - createdAt
    CreateAttendanceRecordBody:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        status:
          type: string
      required:
        - employeeId
        - date
        - status
    PayrollRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        month:
          type: string
        grossPay:
          type: number
        deductions:
          type: number
        netPay:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - employeeName
        - month
        - grossPay
        - deductions
        - netPay
        - status
        - createdAt
    CreatePayrollRecordBody:
      type: object
      properties:
        employeeId:
          type: integer
        month:
          type: string
        grossPay:
          type: number
        deductions:
          type: number
        netPay:
          type: number
        status:
          type: string
      required:
        - employeeId
        - month
        - grossPay
        - deductions
        - netPay
        - status
...