# lib/api-zod/src/generated/types/createPayrollRecordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildRMC Fortune Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePayrollRecordBody {
  employeeId: number;
  month: string;
  grossPay: number;
  deductions: number;
  netPay: number;
  status: string;
}